// web/platform-admin-web/src/api/generated/model/getTermsDocumentResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { GetTermsDocumentResponseDtoStatus } from './getTermsDocumentResponseDtoStatus';

export interface GetTermsDocumentResponseDto {
  /** 약관 문서 식별자 */
  id: string;
  /** 약관 코드 */
  code: string;
  /** 약관 제목 */
  title: string;
  /** 필수 동의 여부 */
  required: boolean;
  /** 종료 일시 */
  terminatedAt?: string;
  /** 약관 상태 */
  status: GetTermsDocumentResponseDtoStatus;
  /** 조직 식별자 */
  organization?: string;
}