// web/platform-admin-web/src/api/generated/model/getTermsDocumentResponseDtoStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 약관 상태
 */
export type GetTermsDocumentResponseDtoStatus = typeof GetTermsDocumentResponseDtoStatus[keyof typeof GetTermsDocumentResponseDtoStatus];


export const GetTermsDocumentResponseDtoStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  TERMINATED: 'TERMINATED',
} as const;