// web/platform-admin-web/src/api/generated/model/getTermsDocumentVersionResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { GetTermsDocumentVersionResponseDtoStatus } from './getTermsDocumentVersionResponseDtoStatus';

export interface GetTermsDocumentVersionResponseDto {
  /** 약관 버전 식별자 */
  id: string;
  /** 약관 내용 */
  content: string;
  /** 체크섬 */
  checksum: string;
  /** 버전 상태 */
  status: GetTermsDocumentVersionResponseDtoStatus;
  /** 효력 일시 */
  effectiveAt?: string;
  /** 버전 라벨 */
  label: string;
}