// web/platform-admin-web/src/api/generated/model/getTermsDocumentVersionResponseDtoStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 버전 상태
 */
export type GetTermsDocumentVersionResponseDtoStatus = typeof GetTermsDocumentVersionResponseDtoStatus[keyof typeof GetTermsDocumentVersionResponseDtoStatus];


export const GetTermsDocumentVersionResponseDtoStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
} as const;